requirements.txt
Django>=1.7.2
dj-static==0.0.6
django-debug-toolbar==1.2.2
django-secure==1.0.1
psycopg2==2.5.4
pytz==2014.10
sqlparse==0.1.13
waitress==0.8.9
newrelic>=2.40
dj-database-url
django_nose==1.3
django-faker==0.2
fake-factory==0.4.2
simplejson==3.6.5